==> R/5_output_stats_and_plots.Rmd <==
---
title:  "Plot CA FRP data"
author: "JJWalker"
date: "2021-09-01"
output:
  html_notebook:
    highlight: tango
    mathjax: null
    number_sections: no
    theme: spacelab
---

```{r, echo = FALSE}

########################################################################### #
#
# 5_output_stats_and_plots.Rmd
#
# Objective:  Generate results (tables and plots) from processed FRP data.
#
# Input:     .RData file produced by 4_calculate_burn_info.Rmd
#
# Output:     FRP plots
#   
# JWalker 7 September 2021
#  
########################################################################### #


# ---- Set environment ----

# Remove existing data and values
#rm(list=ls())

# Load libraries
library(ggplot2)
library(lubridate)
library(kableExtra)
library(knitr)
library(scales)

# Set folder paths
path_in = 'E:/OneDrive - DOI/projects/ca_frp'
path_data <- file.path(path_in, "data")
path_r <- file.path(path_in, 'R')
rdata_out <- "plot_output.Rdata"

path_plots <- file.path(path_in, "/output/plots") 

# Load functions
source(file.path(path_in, "R", "ca_functions.R"))

```
####Data import

```{r, echo = F}

# Load data saved from previous script
rdata_in <- "data_from_script4.Rdata"
load(file = file.path(path_data, rdata_in))

```


```{r, echo = FALSE}

# If the data are not corrected for bowtie effect, slice off extremes

x_fires_evt <- x_fires_evt %>% filter(sample >= 334 & sample <= 1020)

```

```{r, echo = FALSE}



```




###Summary stats
  
```{r, echo = F}

# ---------------------------------------------------------------- #
#### Calculate summary stats ####
# ---------------------------------------------------------------- #
# FRP by burn count (data)
x_frp_summary <- x_fires_evt %>%
                 group_by(burn_num) %>%
                 summarize(meanFRP = mean(frp, na.rm = T),
                           medFRP = median(frp, na.rm = T),
                           n = length(burn_num))
# > x_frp_summary
# # A tibble: 8 x 4
#   burn_num meanFRP medFRP     n
#   <fct>      <dbl>  <dbl> <int>
# 1 1          289.   108.   6594
# 2 2          377.   144.   2131
# 3 3          469.   186.    523
# 4 4          375.   183.    119
# 5 5          423.   280.     39
# 6 6          252.    84.2     4
# 7 7           38.2   38.2     1
# 8 8           10.1   10.1     1

```

####Overall FRP summary

`r x_frp_summary %>%
      kable(digits = 1) %>%
      kable_styling(bootstrap_options = c("striped", "hover"), full_width = F)`

```{r, echo=FALSE}

# Get # of EVT classes with majorities
x_maxevt_gp <- x_fires_evt %>% 
                  group_by(evt_group, burn_num) %>%  #evt_group
                  summarize (meanFRP = mean(frp, na.rm = T), 
                            medFRP = median(frp, na.rm = T), 
                            n = length(burn_num))
# > as.data.frame(x_maxevt_gp)
#      evt_group burn_num   meanFRP medFRP     n
# 1  Agriculture        1       NaN     NA    27
# 2  Agriculture        2  62.40000  45.60    27
# 3  Agriculture        3  78.75000  78.75     2
# 4        Aspen        1 224.40000 224.40     8
# 5        Aspen        2  99.28571  22.20     7
# 6       Barren        1 216.88667 142.35    43
# 7       Barren        2 126.98462  88.10    13
# 8   BirchAspen        1 321.19057 165.00  9332
# 9   BirchAspen        2 243.01473 123.20  4679
# 10  BirchAspen        3 285.64657 111.60   571
# 11  BirchAspen        4 337.02281 109.60    59
# 12  BirchAspen        5  51.53333  58.70     9
# 13  BirchAspen        6  92.46667 100.50     6
# 14  BirchAspen        7 132.45000 132.45     2
# 15  BirchAspen        8  12.60000  12.60     2
# 16   BlkSpruce        1 415.16508 195.00 13918
# 17   BlkSpruce        2 314.56620 149.30  3207
# 18   BlkSpruce        3 271.44379 158.30   169
# 19   BlkSpruce        4  63.40000  63.40     1
# 20      Burned        1 133.59231 122.00    69
# 21      Burned        2 137.25333  68.50    57
# 22      Burned        3  64.09375  47.90    32
# 23      Burned        4   6.10000   6.10     2
# 24  Floodplain        1 223.09853 114.30  4276
# 25  Floodplain        2 208.04540  92.40  1504
# 26  Floodplain        3 155.82308  70.40    91
# 27  Floodplain        4  44.10000  44.10     2
# 28   Grassland        1 426.43622 219.55   735
# 29   Grassland        2 196.87273 116.30   133
# 30   Grassland        3 277.10526 137.30    19
# 31     Hemlock        1 394.35000 394.35     2
# 32       Marsh        1  74.60000  74.60     1
# 33    Peatland        1 161.15177  95.50  1239
# 34    Peatland        2 155.48267  62.20   284
# 35    Peatland        3 111.92500  16.10     8
# 36   Shrubland        1 292.93765 140.95  8030
# 37   Shrubland        2 170.42466  94.00  2948
# 38   Shrubland        3 139.88000  82.35   399
# 39   Shrubland        4  77.01707  69.30    41
# 40   Shrubland        5  92.78889  94.60     9
# 41   Shrubland        6  54.50000  26.40     3
# 42   Shrubland        7  55.00000  55.00     1
# 43   Shrubland        8  18.00000  18.00     1
# 44        Snow        1 142.61250 159.00    19
# 45        Snow        2 456.81818 273.70    11
# 46      Spruce        1 183.35000 142.40     6
# 47       Swamp        1 160.11429  81.50    75
# 48       Swamp        2 106.59200  50.00    29
# 49       Swamp        3  75.50000  70.40     8
# 50       Swamp        4 109.00000 109.00     3
# 51      Tundra        1 235.87352 122.95 10167
# 52      Tundra        2 186.83202  99.40  4204
# 53      Tundra        3 159.45406  80.25   672
# 54      Tundra        4  70.03883  63.00   105
# 55      Tundra        5  76.75306  48.90    49
# 56      Tundra        6  56.43182  40.40    22
# 57      Tundra        7  44.15000  32.05     4
# 58      Tundra        8  40.60000  40.60     1
# 59       Water        1 153.62766  91.60   221
# 60       Water        2 216.11852 124.00    83
# 61       Water        3 338.84000 104.90     5
# 62     Wetland        1 311.65947 148.90 21563
# 63     Wetland        2 238.32670 107.90  5357
# 64     Wetland        3 195.69537 109.10   493
# 65     Wetland        4 246.77636 152.00    55
# 66     Wetland        5  69.50000  51.85    10
# 67     Wetland        6  75.05000  54.70     6
# 68     Wetland        7  32.05000  32.05     2
# 69   WhtSpruce        1 415.42280 199.70 17691
# 70   WhtSpruce        2 294.40111 138.80  4424
# 71   WhtSpruce        3 383.90184 197.60   327
# 72   WhtSpruce        4 143.13333 150.00     9

# Keep only those levels that have at least 10 points
x_maxevt_gp_top <- x_maxevt_gp[x_maxevt_gp$n > 10, ]

# Keep only those classes that have at least 2 burn levels
x_maxevt_gp_top <- x_maxevt_gp_top[!(as.numeric(x_maxevt_gp_top$evt_group) %in% which(table(x_maxevt_gp_top$evt_group) < 2)), ]

```

####FRP summary by primary EVT class

`r x_maxevt_gp_top %>%
      kable(digits = 1) %>%
      kable_styling(bootstrap_options = c("striped", "hover", "condensed"), full_width = F)`
      

```{r, echo=FALSE}

# Get # of EVT classes with majorities. Here use finer distinctions between groups.
x_maxevt_gp_desc <- x_fires_evt %>% 
                  group_by(evt_group_desc, burn_num) %>%  #evt_group
                  summarize (meanFRP = mean(frp, na.rm = T), 
                            medFRP = median(frp, na.rm = T), 
                            n = length(burn_num))


# Keep only those levels that have at least 10 points
x_maxevt_gp_desc_top <- x_maxevt_gp_desc[x_maxevt_gp_desc$n > 10, ]

# Keep only those classes that have at least 2 burn levels
x_maxevt_gp_desc_top <- x_maxevt_gp_desc_top[!(as.numeric(x_maxevt_gp_desc_top$evt_group_desc) %in% which(table(x_maxevt_gp_desc_top$evt_group_desc) < 2)), ]

```     



```{r, echo = FALSE}

# Calculate area of fire pixels per year (in hectares)
x_area_summary <- x_fires_evt %>% 
                    group_by(fire_year) %>% 
                    filter(fire_year > 2001) %>% 
                    summarize(sum_area = sum(evt_sum)*1e-4)  # 1 m^2 = 1e-4
# > x_area_summary
# # A tibble: 19 x 2
#    fire_year sum_area
#        <dbl>    <dbl>
#  1      2002   18462.
#  2      2003    8762.
#  3      2004    5410.
#  4      2005   11162.
#  5      2006   31520.
#  6      2007   21643.
#  7      2008   71236.
#  8      2009   24213.
#  9      2010    2060.
# 10      2011    8155.
# 11      2012   43227.
# 12      2013   26797.
# 13      2014   33776.
# 14      2015   52839.
# 15      2016   29708.
# 16      2017   55336.
# 17      2018   90485.
# 18      2019    6186.
# 19      2020  267580.

```



####FRP summary of fire pixel area detected per year (in hectares)

`r x_area_summary %>%
      kable(digits = 1) %>%
      kable_styling(bootstrap_options = c("striped", "hover", "condensed"), full_width = F)`

```{r, echo= FALSE}

# Retain only the 1st fire in the MODIS era to ensure the veg type is applicable
x_temp1 <- x_fires_evt %>% 
  filter(fire_year > 2001) %>% 
  group_by(ptid) %>% 
  arrange(fire_year) %>% 
  slice(1) %>% 
  ungroup()

x_temp2 <- x_fires_evt %>% 
  filter(fire_year <= 2001)
           
           
x_fires_evt_1stfrp <- rbind(x_temp1, x_temp2)

```

<br>
Data table is saved in this format:
`r x_fires_evt %>%
  arrange(fire_year) %>% 
  filter(ptid == 5767) %>% 
  kable() %>%
  kable_styling()`
  
```{r, echo = FALSE}

# ------------------------------------ -
# #### Set plot parameters ####
# ------------------------------------

# Set consistent plot parameters
dodge = position_dodge(0.16)
theme_set(theme_bw())
plot_opts <- theme(panel.grid.minor.x = element_blank(),
                   panel.grid.major.x = element_blank(), # hide gridlines
                   legend.key = element_blank(),  # remove boxes around legend items
                   plot.title = element_text(hjust = 0.5)) # center title

```


```{r, echo = FALSE, warning=FALSE, message=FALSE}

# Bin FRP into 50 MW bins, plot against frequency of pixels (area) on log scale for all years--
# determine if something is changing wrt the intensity of fires over time
x_sub <- subset(x_fires_evt, !is.na(frp))
ggplot(x_sub, aes(frp)) + geom_histogram() + xlim(0, 3000)


#2.  Winnow those pixels down to those where EVT takes up 90% of the pixel (i.e., "pure" pixels)
#3. Idea: take ALL FRP values for a given pixel, look at spread


```



Boxplot by burn number

```{r, echo= FALSE, warning=FALSE}

# ------- BOXPLOT by burn number, log plot

title <- "Log FRP by burn number, all years"
plot_name <- "CA_FRP by burn number_all years_log.png"


# Plot boxplot with mean values 
p <- ggplot(x_fires_evt, aes(as.factor(burn_num), frp)) +
  scale_y_log10() +
  geom_boxplot(aes(fill = as.factor(burn_num)), show.legend = F) +
  geom_point(data = x_frp_summary, aes(as.factor(burn_num), meanFRP, shape = "mean"), size = 7) +
  scale_shape_manual("", values = 18) +
  geom_text(data = x_frp_summary, aes(label = round(meanFRP, 1), y = meanFRP + 0.7), 
            hjust = -0.1, vjust = -1.1) +
  labs(x = "Burn number", y = "FRP") +
  ggtitle(title)

p + plot_opts

#ggsave(filename = plot_name, path = path_plots, width = 10, height = 7, units = c("in"), dpi = 600)

```


Facet plot of all classes with more than 10 points per burn number

```{r, echo=FALSE, warning=FALSE}

#--------- BOXPLOT by max vegetation classes

x_sub <- subset(x_fires_evt, (evt_group %in% x_maxevt_gp_top$evt_group))

p <- ggplot(x_sub, aes(burn_num, frp)) + 
      geom_boxplot() + 
      scale_y_log10() +
      facet_wrap( ~ evt_group, ncol = 2) #+ 
#      coord_cartesian(ylim = c(0, 800), expand = TRUE)
p + plot_opts

```

Facet plot of all classes (group + desc) with more than 10 points per burn number

```{r, echo=FALSE, warning=FALSE}

#--------- BOXPLOT by max vegetation classes

x_sub <- subset(x_fires_evt, (evt_group_desc %in% x_maxevt_gp_desc_top$evt_group_desc))

p <- ggplot(x_sub, aes(burn_num, frp)) + 
      geom_boxplot() + 
      scale_y_log10() +
      facet_wrap( ~ evt_group_desc, ncol = 2) #+ 
#      coord_cartesian(ylim = c(0, 800), expand = TRUE)
p + plot_opts

```
Restricting the classes to only those of interest gives the following

```{r, echo=FALSE, warning=FALSE}
#--------- BOXPLOT by vegetation groups of interest

goi <- c('Agriculture', 'Barren', 'Developed', 'Forest', 'Grassland',  'Shrubland', 'Water', 'Wetland')
goi <- c('Forest', 'Shrubland', 'Grassland')

x_sub <- subset(x_fires_evt, evt_group %in% goi & as.numeric(burn_num) < 8)
#x_sub <- x_fires_evt

p <- ggplot(x_sub, aes(burn_num, frp)) + 
      geom_boxplot() + 
      scale_y_log10() +
      facet_wrap( ~ evt_group, ncol = 2) #+ 
#      coord_cartesian(ylim = c(0, 800), expand = TRUE)
p + plot_opts

```



```{r, echo=FALSE, warning=FALSE}

#--------- BOXPLOT by vegetation groups NOT of interest

x_sub <- subset(x_fires_evt, !(evt_group %in% goi))

p <- ggplot(x_sub, aes(burn_num, frp)) + 
      geom_boxplot() + 
      scale_y_log10() +
      facet_wrap( ~ evt_group, ncol = 2) #+ 
#      coord_cartesian(ylim = c(0, 800), expand = TRUE)
p + plot_opts

```

Histograms of burn distributions with at least 10 points:
```{r, echo=FALSE, warning=FALSE, message=FALSE}

# -------- HISTOGRAM of distributions

x_sub <- subset(x_fires_evt, as.numeric(burn_num) < 7)
p <- ggplot(x_sub, aes(x=frp, fill = burn_num)) + 
      scale_x_log10() +
      scale_y_log10() +
      geom_histogram(alpha = 0.5, position = "identity") 

p + plot_opts
```

Plots of the effect of time since fire on FRP for all pixels regardless of proportion of EVT class:


```{r, echo=FALSE}
# Limit plot to classes of interest
x_sub <- subset(x_fires_evt) #, evt_group %in% goi)

x_ints <- 
   x_sub %>% filter(reburn == 1) %>% 
   mutate(ints = cut(year_int, c(seq(0, 70, 10)))) %>% 
   group_by(evt_group, ints) %>% 
   summarise(mean_frp = mean(frp, na.rm = T), 
             med_frp = median(frp, na.rm = T),
             n = length(frp))
```


```{r, echo=FALSE}

# Limit plot to classes of interest
x_sub <- subset(x_fires_evt, evt_group %in% goi)

x_ints <- 
   x_sub %>% filter(reburn == 1) %>% 
   mutate(ints = cut(year_int, c(seq(0, 130, 10)))) %>% 
   group_by(evt_group, ints) %>% 
   summarise(mean.frp = mean(frp, na.rm = T), 
             med.frp = median(frp, na.rm = T),
             n = length(frp),
             sd = sd(frp,na.rm = T)) %>% 
             filter(n > 10)

ggplot(x_ints, aes(ints, med.frp, group = evt_group, color = evt_group)) + 
  geom_line() + 
  geom_point() +
  xlab('Years between fires') +
  ylab('Median FRP')
```

```{r, echo=FALSE}

# Limit plot to classes of interest PLUS descriptions
x_sub <- subset(x_fires_evt, evt_group %in% c("Forest")) #, "Shrubland")) 

x_ints <- 
   x_sub %>% filter(reburn == 1) %>% 
   mutate(ints = cut(year_int, c(seq(0, 130, 10)))) %>% 
   group_by(evt_group_desc, ints) %>% 
   summarise(mean_frp = mean(frp, na.rm = T), 
             med_frp = median(frp, na.rm = T),
             n = length(frp),
             sd = sd(frp,na.rm = T)) %>% 
             filter(n > 15)

ggplot(x_ints, aes(ints, mean_frp, group = evt_group_desc, color = evt_group_desc)) + 
  geom_line() + 
  geom_point() +
  xlab('Years between fires') +
  ylab('Median FRP')
```
Look at plots of different proportions of EVT class.  If proportion of max EVT class is at least 20% (and that interval contains at least 10 points):
```{r}

# Set threshold for EVT proportion
evt_min <- 0.20
x_sub <- subset(x_fires_evt, max_evt_prop > evt_min & evt_group %in% goi) 

# Group dataset into time-since-fire intervals
x_ints <- 
   x_sub %>% filter(reburn == 1) %>% 
   mutate(ints = cut(year_int, c(seq(0, 130, by=10)))) %>% 
   group_by(evt_group, ints) %>% 
   summarise(mean_frp = mean(frp, na.rm = T), 
             med_frp = median(frp, na.rm = T),
             n = length(frp),
             sd = sd(frp, na.rm = T)) %>% 
             filter(n > 10) %>% 
             filter(!is.na(ints))

# Plot
ggplot(x_ints, aes(ints, med_frp, group = evt_group, color = evt_group)) + 
  geom_line() + 
  geom_point() +
  xlab('Years between fires') +
  ylab('Median FRP')
```

```{r, echo=FALSE}

# Plot
ggplot(x_ints, aes(ints, mean_frp, group = evt_group, color = evt_group)) + 
  geom_line() + 
  geom_point() +
  xlab('Years between fires') +
  ylab('Mean FRP')


```


If proportion of max EVT class is at least 30%:

```{r}

# Set threshold for EVT proportion
evt_min <- 0.30
x_sub <- subset(x_fires_evt, evt_group %in% goi & max_evt_prop > evt_min)

# Group dataset into time-since-fire intervals
x_ints <- 
   x_sub %>% filter(reburn == 1) %>% 
   mutate(ints = cut(year_int, c(seq(0, 70, 10)))) %>% 
   group_by(evt_group, ints) %>% 
   summarise(mean_frp = mean(frp, na.rm = T), 
             med_frp = median(frp, na.rm = T),
             n = length(frp),
             sd = sd(frp, na.rm = T)) %>% 
             filter(n > 10)  %>% 
             filter(!is.na(ints))

# Plot
ggplot(x_ints, aes(ints, med_frp, group = evt_group, color = evt_group)) + 
  geom_line() + 
  geom_point() +
  xlab('Years between fires') +
  ylab('Median FRP')
```

If proportion of max EVT class is at least 40%:

```{r}

# Set threshold for EVT proportion
evt_min <- 0.40
x_sub <- subset(x_fires_evt, evt_group %in% goi & max_evt_prop > evt_min)

# Group dataset into time-since-fire intervals
x_ints <- 
   x_sub %>% filter(reburn == 1) %>% 
   mutate(ints = cut(year_int, c(seq(0, 70, 10)))) %>% 
   group_by(evt_group, ints) %>% 
   summarise(mean_frp = mean(frp, na.rm = T), 
             med_frp = median(frp, na.rm = T),
             n = length(frp),
             sd = sd(frp, na.rm = T)) %>% 
             filter(n > 10)  %>% 
             filter(!is.na(ints))

# Plot
ggplot(x_ints, aes(ints, med_frp, group = evt_group, color = evt_group)) + 
  geom_line() + 
  geom_point() +
  xlab('Years between fires') +
  ylab('Median FRP')
```

If proportion of max EVT class is at least 50%:

```{r}

# Set threshold for EVT proportion
evt_min <- 0.50
x_sub <- subset(x_fires_evt, evt_group %in% goi & max_evt_prop > evt_min)

# Group dataset into time-since-fire intervals
x_ints <- 
   x_sub %>% filter(reburn == 1) %>% 
   mutate(ints = cut(year_int, c(seq(0, 70, 10)))) %>% 
   group_by(evt_group, ints) %>% 
   summarise(mean_frp = mean(frp, na.rm = T), 
             med_frp = median(frp, na.rm = T),
             n = length(frp),
             sd = sd(frp, na.rm = T)) %>% 
             filter(n > 10)  %>% 
             filter(!is.na(ints))

# Plot
ggplot(x_ints, aes(ints, med_frp, group = evt_group, color = evt_group)) + 
  geom_line() + 
  geom_point() +
  xlab('Years between fires') +
  ylab('Median FRP')


```
If proportion of max EVT class is at least 60%:


```{r}

# Set threshold for EVT proportion
evt_min <- 0.60
x_sub <- subset(x_fires_evt, evt_group %in% goi & max_evt_prop > evt_min)

# Group dataset into time-since-fire intervals
x_ints <- 
   x_sub %>% filter(reburn == 1) %>% 
   mutate(ints = cut(year_int, c(seq(0, 70, 10)))) %>% 
   group_by(evt_group, ints) %>% 
   summarise(mean_frp = mean(frp, na.rm = T), 
             med_frp = median(frp, na.rm = T),
             n = length(frp),
             sd = sd(frp, na.rm = T)) %>% 
             filter(n > 10)  %>% 
             filter(!is.na(ints))

# Plot
ggplot(x_ints, aes(ints, med_frp, group = evt_group, color = evt_group)) + 
  geom_line() + 
  geom_point() +
  xlab('Years between fires') +
  ylab('Median FRP')


```




```{r}

# Set threshold for EVT proportion
evt_min <- 0.70
x_sub <- subset(x_fires_evt, evt_group %in% goi & max_evt_prop > evt_min)

# Group dataset into time-since-fire intervals
x_ints <- 
   x_sub %>% filter(reburn == 1) %>% 
   mutate(ints = cut(year_int, c(seq(0, 70, 10)))) %>% 
   group_by(evt_group, ints) %>% 
   summarise(mean_frp = mean(frp, na.rm = T), 
             med_frp = median(frp, na.rm = T),
             n = length(frp),
             sd = sd(frp, na.rm = T)) %>% 
             filter(n > 10)  %>% 
             filter(!is.na(ints))

# Plot
ggplot(x_ints, aes(ints, med_frp, group = evt_group, color = evt_group)) + 
  geom_line() + 
  geom_point() +
  xlab('Years between fires') +
  ylab('Median FRP')


```




```{r, echo=F}

x_sub$bins <- cut(x_sub$year_int, breaks=c(seq(0, 100, 10)), labels=c("1-10","10-20","20-30", '30-40', '40-50', '50-60', '60-70', '70-80', '80-90', '90-100'))
                   




```


```{r, echo=FALSE}
# For individual years. This mostly groups data geographically as well.

for (yr in seq(2014, 2020)) {
  
  x_yr <- subset(x_fires_evt, fire_year == yr)
  
# ------- PLOT by burn number
  title <- paste0("FRP by burn number, ", yr)
  plot_name <- paste0("AK_FRP by burn number_", yr, ".png")

  p <- ggplot(x_yr, aes(as.factor(burn_num), log10(frp))) + 
    geom_boxplot() + 
    labs(x = "Burn number", y = "log FRP") +
    ggtitle(title)
  p + plot_opts
  ggsave(filename = plot_name, path = path_plots, width = 10, height = 7, units = c("in"), dpi = 600)


# ------ PLOT by years since prior burn
# Drop factors for which there are fewer than 10 points
  keep <- levels(as.factor(x_yr$year_int))[table(x_yr$year_int) > 10]
  x_sub_int <- x_yr[x_yr$year_int %in% keep, ]

  title <- paste0("FRP by years since prior burn, ", yr)
  plot_name <- paste0("AK_FRP by years since prior burn_", yr, ".png")

  p <- ggplot(x_sub_int, aes(as.factor(year_int), log10(frp))) + 
       geom_boxplot() + 
       #ylim(0, 1000) + 
       labs(x = "Years since prior burn", y = "log FRP") +
       ggtitle(title)
  p + plot_opts

  ggsave(filename = plot_name, path = path_plots, width = 10, height = 7, units = c("in"), dpi = 600)

# ------- PLOT by year of fire
  keep <- levels(as.factor(x_yr$FireYear))[table(x_yr$fire_year) > 10]
  x_sub_yr <- x_yr[x_yr$fire_year %in% keep, ]

  title <- paste0("FRP by year of fire, ", yr)
  plot_name <- paste0("AK_FRP by year of fire_", yr, ".png")

  p <- ggplot(x_sub_yr, aes(as.factor(fire_year), log10(frp))) + 
       geom_boxplot() + 
       #ylim(0, 1000) +
       labs(x = "Year of fire", y = "log FRP") +
       ggtitle(title)
  p + plot_opts

  ggsave(filename = plot_name, path = path_plots, width = 10, height = 7, units = c("in"), dpi = 600)
  
}
```

```{r, echo = FALSE}

# -------- PLOT by mean fire interval
#  x_pts <- x_sub_int %>% group_by(year_int) %>% summarize(avg = mean(frp))

#x_pts <- subset(x, evt_group == 'WhtSpruce')
x_pts <-  x_sub_int %>% group_by(burn_num, year_int) %>% summarise(avg = mean(frp))

lm_eqn2 = function(x){
  m = lm(y ~ poly(x, 3), x)#3rd degree polynomial
  eq <- substitute(italic(y) == a + b %.% italic(x)*","~~italic(r)^2~"="~r2,
                   list(a = format(coef(m)[1], digits = 2),
                        b = format(coef(m)[2], digits = 2),
                        r2 = format(summary(m)$r.squared, digits = 3)))
  as.character(as.expression(eq))
}

  #title <- paste0("Mean FRP by year interval, ", yr)
  #plot_name <- paste0("AK_FRP mean FRP by year ", yr, ".png")

  p <- ggplot(x_pts, aes(year_int, avg)) + 
       geom_point() + 
       geom_smooth(method = 'lm', se = FALSE, size = 0.4) +
       labs(x = "Years since prior burn", y = "FRP") +
       annotate("text", x = 5, y = 400, label = lm_eqn2(x_pts, "avg", "year_int"), 
           parse = TRUE, hjust = 0, vjust = 0, color = "#000000") +
       ggtitle(title)
  p + plot_opts

#ggsave(filename = plot_name, path = path_plots, width = 10, height = 7, units = c("in"), dpi = 600)
```

# Calculate semivariograms
x_copy <- x_yr

# convert degrees to km distances
x_copy$lon_km <- 110. * (x_copy$longitude - (-125))*cos(x_copy$latitude/(360/(2*pi)))
x_copy$lat_km <- 110. * (x_copy$latitude-65)
coordinates(x_copy) = ~lon_km+lat_km


# ------- PLOT points in x,y km space  
plot_name <- paste0("AK_FRP bubble plot ", yr, ".png")
png(file = file.path(path_plots, plot_name), width = 10, height = 7, units = c("in"), res = 600)
print(bubble(x_copy, zcol = 'MaxFRP', fill = T, do.sqrt = F, maxsize = 1.8))
dev.off()

# ------- PLOT variogram
v = variogram(MaxFRP~1, x_copy)
level <- c(0.5, 0.95)

plot_name <- paste0("AK_FRP variogram ", yr, ".png")
png(file = file.path(path_plots, plot_name), width = 10, height = 7, units = c("in"), res = 600)
print(plot(v, fraction = 0.65, level=level))
dev.off()


# # ------PLOT directional variogram
#   v.fit = fit.variogram(v, vgm("Sph"))
#   x_copy.var.det <- variogram(MaxFRP~lat_km+lon_km, data = x_copy)
#   
#   plot_name <- paste0("AK_FRP directional variogram ", yr, ".png")
#   p <- plot(x_copy.var.det)
#   p

plot_name <- paste0("AK_FRP directional variogram ", yr, ".png")

x_copy.var.dir <- variogram(MaxFRP~1, data = x_copy, alpha = c(0, 45, 90, 135))
png(file = file.path(path_plots, plot_name), width = 10, height = 7, units = c("in"), res = 600)
print(plot(x_copy.var.dir))
dev.off()

}


```{r, echo = FALSE}

# --------------------------------- 
# Stats
# ---------------------------------

# Calculate 2-way ANOVA; evt group and burn num
x_anova <- aov(formula = frp ~ evt_group * burn_num, data = x_fires_evt)
summary(x_anova)
TukeyHSD(x_anova, which = "burn_num")
TukeyHSD(x_anova, which = "evt_group")


# Calculate anova: FRP vs year_int, evt, and burn_num
x_anova <- aov(formula = frp ~ year_int * as.factor(evt_group) * burn_num, data = subset(x_fires_evt, reburn > 0))
summary(x_anova)

```



Frequency plot by year
```{r, echo=FALSE, warning=FALSE}

# Create frequency tables
x_cut <- x_fires_evt[complete.cases(x_fires_evt[, 'frp']), ]
x_cut$bin <- cut(x_cut$frp, breaks = c(seq(0, 12000, 100)), include.lowest = T)

x_year_bin <- x_cut %>% 
          group_by(bin, fire_year) %>% 
          count()

# Break out the lower and upper ranges of the binned intervals
x_year_bin <- cbind(x_year_bin, lower = as.numeric( sub("\\((.+),.*", "\\1", x_year_bin$bin) ),
      upper = as.numeric( sub("[^,]*,([^]]*)\\]", "\\1", x_year_bin$bin) ))

x.year.prop <- x_year_bin %>%
  group_by(fire_year) %>% 
  mutate(prop = n/sum(n)) 

ggplot(x.year.prop, aes(upper, prop, group = as.factor(fire_year), 
                        color = as.factor(fire_year),
                        shape = as.factor(fire_year))) + 
  geom_point()  + 
  theme_bw() +
  xlab("FRP") +
  ylab('Frequency') +
  scale_x_log10() + 
  scale_y_log10() +
  scale_shape_manual(name = 'Year', values=c(19, 3, 19, 4, 19, 5, 19, 6, 19, 7, 19, 8, 19, 9, 20, 19, 8, 19, 9, 20)) +
  scale_color_discrete(name = 'Year') +
  annotation_logticks(scaled = TRUE)



```

Frequency plot by max veg class

```{r, echo=FALSE, warning=FALSE}

goi <- c('Developed', 'Forest', 'Grassland',  'Shrubland')

# Create frequency tables
x_evt_sub <- x_fires_evt[x_fires_evt$evt_group %in% goi & x_fires_evt$frp < 10000, ]
x_evt_cut <- x_evt_sub[complete.cases(x_evt_sub[, 'frp']), ]
x_evt_cut$bin <- cut(x_evt_cut$frp, breaks = c(seq(0, 12000, 100)), include.lowest = T)

x_evt_bin <- x_evt_cut %>% 
          group_by(bin, evt_group) %>% 
          count()

# Break out the lower and upper ranges of the binned intervals
x_cut_evt_bin <- cbind(x_evt_bin, lower = as.numeric( sub("\\((.+),.*", "\\1", x_evt_bin$bin) ),
      upper = as.numeric( sub("[^,]*,([^]]*)\\]", "\\1", x_evt_bin$bin) ))


x_evt_prop <- x_cut_evt_bin %>%
  group_by(evt_group) %>% 
  mutate(prop = n/sum(n)) 


ggplot(subset(x_evt_prop, evt_group %in% goi), aes(upper, prop, group = evt_group, color = evt_group, shape = evt_group)) + 
  geom_point() +
  scale_x_log10() + 
  scale_y_log10(limits = c(1e-04, 1)) +
  xlab("FRP") +
  ylab('Frequency') +
  scale_shape_manual(name = 'EVT group', values=c(19, 3, 19, 4, 19, 5, 19, 6),
                     breaks=c("Developed", "Forest", "Grassland", "Shrubland"),
                       labels=c( "Developed", "Forest", "Grassland", "Shrubland")) +
  scale_color_discrete(name = "EVT group",
                       breaks=c( "Developed", "Forest", "Grassland", "Shrubland"),
                       labels=c( "Developed", "Forest", "Grassland", "Shrubland")) +
  annotation_logticks(scaled = TRUE)


```
